More osc

+++ src/voice.rs
@@ -75,7 +75,7 @@ impl<S: Oscillator> Voice<S> {
         self.phase +=
             ((self.phase >= 1.0) as u8 as f64 * -1.0) + ((self.phase < 0.0) as u8 as f64 * 1.0);
 
-        self.signal.osc(self.phase)
+        self.signal.osc(self.phase) as f32
     }
 
     pub(crate) fn set_samplerate(&mut self, sr: f32) {
